Move responsibilities up a layer
With the creation of a SecretsManager client we want to move some
responsibilities up.

Moving the retrieval of the Google Calendar ID and the Google Calendar
service account offloads those responsibilities from the emitter, which
shouldn't need to care about those things.

We move all that to a top level main function which uses our
new/adjusted components.

This also paves the way for making these components more reusable for
future ideas of scraping and emitting different calendars.