Remove spacing around calendar entries
I originally assumed that the calendar entries on the open field hours
website were always generated in a uniform fashion. This assumption was
broken recently by a space between a time and the dash in one of the
calendar entry time blocks. This seems to be a data entry typo.

In the future there could be instances of other typos in these website
calendar entries. In anticipation, I've created a function handling
sanitizing these calendar entries. For now, it only handles the most
recent issue of spacing between the dashes.