Restructure project for pytest
Introducing pytest into the project for when I eventually introduce
circleci. For that I'll want to run all the tests simultaneously. Pytest
will let me do that but in order to get it working I needed to
restructure the project files into a source directory and adjust the
import statements to match.