Add an AWS CLI install step in the CircleCI config
We want to specify a specific Python version to be used as the base
docker image. But, the base image doesn't come with AWS CLI installed,
which we need to push the built image to ECR and update Lambda.

We could have either used a base AWS CLI image and installed Python, but
that requires more steps than installing AWS CLI.

We could have also updated our Dockerfile to create a custom image
defining the Python version and installing the AWS CLI, then use that
as our CircleCI image. However, we don't need AWS CLI in our image of
the project so installing in CircleCI separates the concerns.