Remove unnecessary scraper daylight savings test
I should have expected this test to be unnecessary. Since daylight
savings time jumps happen early morning like 2 AM, the open field hours
schedule is unaffected. The website treats it as any other day and the
time block looks the same as others.

I thought about leaving this test in for the sake of documenting the
case. I'd rather just leave the test out and not have to duplicate the
successful case test. It was good to put it in at the start but I think
it's prudent to remove it now that it's confirmed irrelevant.