Remove unnecessary scraper new year test
This test ended up being unnecessary because the schedule on the website
worked for the new year the same as any new month. I was originally
thinking that there'd be some special calendar across both the end of
December and the start of January. The calendar just ended up being all
of December and then the first half of January, with no month/year
overlap.

It was a good idea to jot down this possible case, but we don't
need it anymore now that we know the behavior. Removal of the test case
rather than keeping it and duplicating the typical successful case
follows the same logic as the previous commit removing the unnecessary
daylight savings test.