Fix cases testing for expected datetime in scraper
The successful field hours case and the typo case both had tests
expecting the tested function to return datetimes in 2024. However, we
expect the datetime tuples being extracted from the test html and from
the actual Asphalt Green website to just specify the month and day. Thus
the expected datetimes should be whatever year it currently is, which is
currently 2025. So, we change our expected datetime tuples accordingly
(use current year rather than a fixed value).